+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
  - [Fix] Check if payment transaction is already sent to TipiRégie webservice and recreate it in this case.
+ - [Fix] Set real validation datetime from TipiRégie webservice information
 
 ## [10.0.1.3.0] - 2019-07-19
 ### Added